requirements.txt
openai
langchain
langchain-community
langchain-openai
autogen-py
tiktoken
faiss-cpu
python-dotenv
pyyaml
